feature/db6-01-extract-shared-sql-constants
