+++ src/features/bed-dashboard/lib/discharge-queries.ts
@@ -4,30 +4,7 @@
 // Extracted from discharge-actions.ts to stay under the 200-line file limit.
 
 import type { PoolClient } from 'pg'
-
-/**
- * US-8.2: Inline subquery that resolves the active shift for the current
- * wall-clock time. Handles the Night shift that crosses midnight
- * (start_time > end_time). Identical pattern to INSERT_BED_STAGE_LOG_SQL
- * in bed-mutations.constants.ts — keep both in sync.
- */
-const SHIFT_AUTOTAG_SUBQUERY = `
-  (
-    SELECT s.id
-    FROM   shifts s
-    WHERE  s.is_active = TRUE
-      AND (
-        (s.start_time <= s.end_time
-           AND NOW()::time >= s.start_time
-           AND NOW()::time <  s.end_time)
-        OR
-        (s.start_time > s.end_time
-           AND (NOW()::time >= s.start_time OR NOW()::time < s.end_time))
-      )
-    ORDER BY s.is_default DESC, s.created_at ASC
-    LIMIT 1
-  )
-`
+import { SHIFT_AUTOTAG_SUBQUERY } from '@/shared/lib/shift-helpers'
 
 export interface BedDischargeRow {
   id: string
